+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - #71 - Front CRUD for User
 - #84 - Generate test data for ProductTestControllerIT
 - #83 - Generate test data for PostomatRestControllerIT
+- #12 - Order entity
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity